backend/pyproject.toml
[project]
name = "cloudnature-estimate-backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "sqlmodel>=0.0.22",
    "pydantic-settings>=2.7.0",
    "openai>=1.60.0",
    "resend>=2.0.0",
    "python-dotenv>=1.0.1",
    "httpx>=0.28.0",
    "jsonschema>=4.23.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.25.0",
    "httpx>=0.28.0",
]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.backends._legacy:_Backend"
